Rename InvocablePromise to Promise.Invocable
